Cleaning things up a bit more, fixing InnerError
